Fix: pylint errors and Refactor: better naming
removed some pylint errors/warnings, it's really pedantic and wrongly so
in lots of cases. For instance, all of our Flask decorated functions
are reported as unused, but they are. See https://github.com/PyCQA/pylint/issues/2842 for a discussion on the topic, I get that it's possible that the decorator doesn't actually use the function. In addition renamed routes for culture and admin to have simpler names, removing the preface of `admin_` or `culture_`. Maybe it's wortwhile to move to Flask RESTful now?