Feat: better errors & fallback to downloaded
Errors now display a Snackbar when encountered from Admin POV, for users
when encountering an error when loading information about a culture we
just go back, would be better to display some content, but not bad. When
a user fails to connect to the Api the Ledger will attempt to read local
storage to get information about the culture, if that fails go back.