Fix(frontend): App theming jankiness
In the previous version in order to get the <PaperProvider /> component to
re-render <App /> subscribed to the Redux Store in order to cause a
manual update. This is now avoided by using react-redux connect in order
to bind a new Navigator component to the state of the store causing a
re-render on theme change.